requirements.txt
pygame-ce>=2.4.1
Pillow>=10.0.0
watchdog>=3.0.0
